+++ db.json
@@ -133,27 +133,23 @@
   "dataBasket": [
     {
       "type": "pizza",
-      "id": "p2",
+      "id": "P6",
       "img": "./src/img/887590c4.jpeg",
-      "title": "Salmone",
-      "description": "'томат, моцарела, морепродукти, лимон, sadsadasd asd as пармезан',",
+      "title": "ZCCcasdd",
+      "description": "томат, моцарела, томат, моцарела, пармезан пармезан",
+      "size": "32",
+      "weight": "250",
+      "price": 180
+    },
+    {
+      "type": "pizza",
+      "id": "p5",
+      "img": "./src/img/887590c4.jpeg",
+      "title": "anasda",
+      "description": "томат, моцарела, пармезан",
       "size": "35",
       "weight": "450",
       "price": 220
-    },
-    {
-      "img": "src/img/887590c4.jpeg",
-      "title": "Аллегро",
-      "description": "салат Айсберг, бастурма, перець болгарський, помідори чері, фета, горіхи кеш'ю, соус",
-      "price": 110,
-      "id": "0AJRDp8"
-    },
-    {
-      "img": "src/img/887590c4.jpeg",
-      "title": "Аллегро",
-      "description": "салат Айсберг, бастурма, перець болгарський, помідори чері, фета, горіхи кеш'ю, соус",
-      "price": 110,
-      "id": "UecJT74"
     }
   ]
 }
